+++ lib/effects/effects.dart
@@ -12,6 +12,7 @@ export './sequence_effect.dart';
 
 abstract class PositionComponentEffect {
   PositionComponent component;
+  Function() onComplete;
 
   bool _isDisposed = false;
   bool get isDisposed => _isDisposed;
@@ -37,7 +38,11 @@ abstract class PositionComponentEffect {
   /// travel time
   double get totalTravelTime => travelTime * (isAlternating ? 2 : 1);
 
-  PositionComponentEffect(this._initialIsInfinite, this._initialIsAlternating) {
+  PositionComponentEffect(
+    this._initialIsInfinite,
+    this._initialIsAlternating, {
+    this.onComplete,
+  }) {
     isInfinite = _initialIsInfinite;
     isAlternating = _initialIsAlternating;
   }
@@ -53,6 +58,9 @@ abstract class PositionComponentEffect {
     if (!hasFinished()) {
       currentTime += dt * curveDirection + driftTime * driftMultiplier;
       percentage = min(1.0, max(0.0, currentTime / travelTime));
+      if (hasFinished()) {
+        onComplete?.call();
+      }
     }
   }